requirements.txt
paho-mqtt
streamz
environs